pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "photo-darkroom-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "nicegui~=2.18.0",
    "platformdirs~=4.0.0",
    "pydantic~=2.10.0",
    "pywebview~=6.1.0",
    "pyyaml~=6.0.0",
]

[project.scripts]
photo-darkroom-manager = "photo_darkroom_manager.gui.gui_app:main"
dr-mng = "photo_darkroom_manager.gui.gui_app:main"

[dependency-groups]
dev = ["pre-commit~=4.5.1", "pytest~=8.4.0", "ruff~=0.15.8", "ty~=0.0.28"]

[tool.hatch.build.targets.wheel]
packages = ["src/photo_darkroom_manager"]

[tool.ruff]
target-version = "py314"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]

[tool.ruff.lint.isort]
known-first-party = ["photo_darkroom_manager"]

[tool.pytest.ini_options]
testpaths = ["tests"]
